+++ wampum.php
@@ -16,7 +16,7 @@
  * Text Domain:        wampum
  * License:            GPL-2.0+
  * License URI:        http://www.gnu.org/licenses/gpl-2.0.txt
- * Version:            1.4.3
+ * Version:            1.4.4
  * GitHub Plugin URI:  https://github.com/JiveDig/wampum
  * GitHub Branch:	   master
  */
@@ -165,7 +165,7 @@ final class Wampum_Setup {
 
 		// Plugin version.
 		if ( ! defined( 'WAMPUM_VERSION' ) ) {
-			define( 'WAMPUM_VERSION', '1.4.3' );
+			define( 'WAMPUM_VERSION', '1.4.4' );
 		}
 
 		// Plugin Folder Path.